requirements.txt
streamlit
numpy
opencv-python
pillow
tensorflow
python-dotenv
requests
google-generativeai
reportlab
